+++ src/pools/selection-strategies/selection-strategies-types.ts
@@ -16,8 +16,6 @@ export const WorkerChoiceStrategies = Object.freeze({
   LEAST_BUSY: 'LEAST_BUSY',
   /**
    * Least ELU worker selection strategy.
-   *
-   * @experimental
    */
   LEAST_ELU: 'LEAST_ELU',
   /**