+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix worker choice strategy retries mechanism on some edge cases.
+- Fix worker choice strategy retries mechanism in some edge cases.
 
 ### Changed