+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.31] - 2023-08-20
+
 ### Fixed
 
 - Fix worker choice strategy retries mechanism in some edge cases.